Rename channels.images to rooms.images
