+++ packages/rest-typings/src/v1/rooms.ts
@@ -1,4 +1,4 @@
-import type { IMessage, IRoom, IUser, RoomAdminFieldsType } from '@rocket.chat/core-typings';
+import type { IMessage, IRoom, IUser, RoomAdminFieldsType, IUpload } from '@rocket.chat/core-typings';
 import Ajv from 'ajv';
 
 import type { PaginatedRequest } from '../helpers/PaginatedRequest';
@@ -436,6 +436,15 @@ export type Notifications = {
 
 type RoomsGetDiscussionsProps = PaginatedRequest<BaseRoomsProps>;
 
+export type RoomsImagesProps = {
+	roomId: string;
+	startingFromId: string;
+	count?: number;
+	offset?: number;
+};
+const roomsImagesPropsSchema = {};
+export const isRoomsImagesProps = ajv.compile<RoomsImagesProps>(roomsImagesPropsSchema);
+
 export type RoomsEndpoints = {
 	'/v1/rooms.autocomplete.channelAndPrivate': {
 		GET: (params: RoomsAutoCompleteChannelAndPrivateProps) => {
@@ -571,6 +580,11 @@ export type RoomsEndpoints = {
 	'/v1/rooms.getDiscussions': {
 		GET: (params: RoomsGetDiscussionsProps) => PaginatedResult<{
 			discussions: IRoom[];
+		}>;
+	};
+	'/v1/rooms.images': {
+		GET: (params: RoomsImagesProps) => PaginatedResult<{
+			files: IUpload[];
 		}>;
 	};
 };
